requirements.txt
asgiref==3.9.1
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
class-registry==2.1.2
cryptography==45.0.5
Django>=5.1,<5.2
django-auto-prefetching==0.2.12
django-cgi==0.1.2
django-cors-headers==4.7.0
django-daraja==1.3.0
django-filter==25.1
django-jazzmin==3.0.1
django-requests==0.0.2
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
filters==1.3.2
flake8==3.8.4
idna==3.10
inflection==0.5.1
mccabe==0.6.1
more-itertools==10.7.0
packaging==25.0
pillow==11.3.0
pycodestyle==2.6.0
pycparser==2.22
pyflakes==2.2.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
qrcode==8.2
regex==2025.7.34
reportlab==4.4.3
requests==2.32.4
six==1.17.0
sqlparse==0.3.1
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.9.0
